+++ test/unit/test_dgraph.py
@@ -10,9 +10,9 @@ from tgm.data import DGData
 @pytest.fixture
 def data():
     edge_index = torch.IntTensor([[2, 2], [2, 4], [1, 8]])
-    edge_timestamps = torch.IntTensor([1, 5, 20])
+    edge_timestamps = torch.LongTensor([1, 5, 20])
     edge_feats = torch.rand(3, 5)
-    node_timestamps = torch.IntTensor([1, 5, 10])
+    node_timestamps = torch.LongTensor([1, 5, 10])
     node_ids = torch.IntTensor([2, 4, 6])
     dynamic_node_feats = torch.rand([3, 5])
     static_node_feats = torch.rand(9, 11)
@@ -48,7 +48,7 @@ def test_init_from_data(data):
     expected_edges = (
         torch.tensor([2, 2, 1], dtype=torch.int32),
         torch.tensor([2, 4, 8], dtype=torch.int32),
-        torch.tensor([1, 5, 20], dtype=torch.int32),
+        torch.tensor([1, 5, 20], dtype=torch.int64),
     )
     torch.testing.assert_close(dg.edges, expected_edges)
 
@@ -69,7 +69,7 @@ def test_init_gpu(data):
     expected_edges = (
         torch.tensor([2, 2, 1], dtype=torch.int32, device='cuda'),
         torch.tensor([2, 4, 8], dtype=torch.int32, device='cuda'),
-        torch.tensor([1, 5, 20], dtype=torch.int32, device='cuda'),
+        torch.tensor([1, 5, 20], dtype=torch.int64, device='cuda'),
     )
     torch.testing.assert_close(dg.edges, expected_edges)
 
@@ -104,7 +104,7 @@ def test_materialize(data):
     dg = DGraph(data)
     exp_src = torch.tensor([2, 2, 1], dtype=torch.int32)
     exp_dst = torch.tensor([2, 4, 8], dtype=torch.int32)
-    exp_t = torch.tensor([1, 5, 20], dtype=torch.int32)
+    exp_t = torch.tensor([1, 5, 20], dtype=torch.int64)
     exp = DGBatch(
         exp_src,
         exp_dst,
@@ -121,7 +121,7 @@ def test_materialize_skip_feature_materialization(data):
     dg = DGraph(data)
     exp_src = torch.tensor([2, 2, 1], dtype=torch.int32)
     exp_dst = torch.tensor([2, 4, 8], dtype=torch.int32)
-    exp_t = torch.tensor([1, 5, 20], dtype=torch.int32)
+    exp_t = torch.tensor([1, 5, 20], dtype=torch.int64)
     exp = DGBatch(exp_src, exp_dst, exp_t, None, None)
     torch.testing.assert_close(
         asdict(dg.materialize(materialize_features=False)), asdict(exp)
@@ -152,7 +152,7 @@ def test_slice_time_no_upper_bound(data):
     exp_edges = (
         torch.IntTensor([2, 1]),
         torch.IntTensor([4, 8]),
-        torch.IntTensor([5, 20]),
+        torch.LongTensor([5, 20]),
     )
     torch.testing.assert_close(dg1.edges, exp_edges)
 
@@ -183,7 +183,7 @@ def test_slice_time_at_end_time(data):
     exp_edges = (
         torch.IntTensor([2, 2]),
         torch.IntTensor([2, 4]),
-        torch.IntTensor([1, 5]),
+        torch.LongTensor([1, 5]),
     )
     torch.testing.assert_close(dg1.edges, exp_edges)
 
@@ -223,7 +223,7 @@ def test_slice_time_to_empty(data):
     exp_edges = (
         torch.IntTensor([2, 2]),
         torch.IntTensor([2, 4]),
-        torch.IntTensor([1, 5]),
+        torch.LongTensor([1, 5]),
     )
     torch.testing.assert_close(dg1.edges, exp_edges)
 
@@ -249,7 +249,7 @@ def test_slice_time_to_empty(data):
     exp_edges = (
         torch.IntTensor([2]),
         torch.IntTensor([4]),
-        torch.IntTensor([5]),
+        torch.LongTensor([5]),
     )
     torch.testing.assert_close(dg2.edges, exp_edges)
 
@@ -271,7 +271,7 @@ def test_slice_time_to_empty(data):
     assert dg3.num_timestamps == 1
     assert dg3.nodes == {6}
 
-    exp_edges = (torch.IntTensor([]), torch.IntTensor([]), torch.IntTensor([]))
+    exp_edges = (torch.IntTensor([]), torch.IntTensor([]), torch.LongTensor([]))
     torch.testing.assert_close(dg3.edges, exp_edges)
 
     exp_dynamic_node_feats = torch.zeros(dg3.end_time + 1, dg3.num_nodes, 5)
@@ -294,7 +294,7 @@ def test_slice_time_to_empty(data):
     assert dg4.dynamic_node_feats is None
     assert dg4.edge_feats is None
 
-    exp_edges = (torch.IntTensor([]), torch.IntTensor([]), torch.IntTensor([]))
+    exp_edges = (torch.IntTensor([]), torch.IntTensor([]), torch.LongTensor([]))
     torch.testing.assert_close(dg4.edges, exp_edges)
 
     # Check original graph cache is not updated
@@ -331,7 +331,7 @@ def test_slice_events(data):
     exp_edges = (
         torch.IntTensor([2]),
         torch.IntTensor([4]),
-        torch.IntTensor([5]),
+        torch.LongTensor([5]),
     )
     torch.testing.assert_close(dg1.edges, exp_edges)
 
@@ -375,7 +375,7 @@ def test_slice_events_slice_time_combination(data):
     exp_edges = (
         torch.IntTensor([2]),
         torch.IntTensor([4]),
-        torch.IntTensor([5]),
+        torch.LongTensor([5]),
     )
     torch.testing.assert_close(dg1.edges, exp_edges)
