+++ test/unit/test_split.py
@@ -17,7 +17,7 @@ def test_time_split_bad_args():
 
 
 def test_temporal_split():
-    edge_times = torch.IntTensor([1, 2, 3, 4])
+    edge_times = torch.LongTensor([1, 2, 3, 4])
     edge_index = torch.IntTensor([[1, 2], [3, 4], [5, 6], [7, 8]])
     num_nodes = edge_index.max() + 1
     static_node_feats = torch.rand(num_nodes, 5)
@@ -52,10 +52,10 @@ def test_temporal_split():
 
 
 def test_temporal_split_with_node_feats():
-    edge_times = torch.IntTensor([1, 2, 3, 4])
+    edge_times = torch.LongTensor([1, 2, 3, 4])
     edge_index = torch.IntTensor([[1, 2], [3, 4], [5, 6], [7, 8]])
 
-    node_times = torch.IntTensor([1, 2, 4])
+    node_times = torch.LongTensor([1, 2, 4])
     node_ids = torch.IntTensor([1, 2, 3])
     dynamic_node_feats = torch.rand(3, 7)
     num_nodes = edge_index.max() + 1
@@ -95,10 +95,10 @@ def test_temporal_split_with_node_feats():
 
 
 def test_temporal_split_only_train_split():
-    edge_times = torch.IntTensor([1, 2, 3, 4])
+    edge_times = torch.LongTensor([1, 2, 3, 4])
     edge_index = torch.IntTensor([[1, 2], [3, 4], [5, 6], [7, 8]])
 
-    node_times = torch.IntTensor([0, 2, 4])
+    node_times = torch.LongTensor([0, 2, 4])
     node_ids = torch.IntTensor([1, 2, 3])
     dynamic_node_feats = torch.rand(3, 7)
     num_nodes = edge_index.max() + 1
@@ -129,7 +129,7 @@ def test_temporal_ratio_split_bad_args():
 
 
 def test_temporal_ratio_split():
-    edge_times = torch.IntTensor([1, 2, 3, 4])
+    edge_times = torch.LongTensor([1, 2, 3, 4])
     edge_index = torch.IntTensor([[1, 2], [3, 4], [5, 6], [7, 8]])
     num_nodes = edge_index.max() + 1
     static_node_feats = torch.rand(num_nodes, 5)
@@ -164,10 +164,10 @@ def test_temporal_ratio_split():
 
 
 def test_temporal_ratio_split_with_node_feats():
-    edge_times = torch.IntTensor([1, 2, 3, 4])
+    edge_times = torch.LongTensor([1, 2, 3, 4])
     edge_index = torch.IntTensor([[1, 2], [3, 4], [5, 6], [7, 8]])
 
-    node_times = torch.IntTensor([1, 2, 4])
+    node_times = torch.LongTensor([1, 2, 4])
     node_ids = torch.IntTensor([1, 2, 3])
     dynamic_node_feats = torch.rand(3, 7)
     num_nodes = edge_index.max() + 1
@@ -210,10 +210,10 @@ def test_temporal_ratio_split_with_node_feats():
 
 
 def test_temporal_ratio_split_only_train_split():
-    edge_times = torch.IntTensor([1, 2, 3, 4])
+    edge_times = torch.LongTensor([1, 2, 3, 4])
     edge_index = torch.IntTensor([[1, 2], [3, 4], [5, 6], [7, 8]])
 
-    node_times = torch.IntTensor([0, 2, 4])
+    node_times = torch.LongTensor([0, 2, 4])
     node_ids = torch.IntTensor([1, 2, 3])
     dynamic_node_feats = torch.rand(3, 7)
     num_nodes = edge_index.max() + 1
